SCAT-1103 Написать тест на отключенную галочку "Запомнить меня"

+++ AbbyyLS.SmartCAT.Selenium.Tests/Drivers/ChromeDriverProvider.cs
@@ -10,7 +10,7 @@ namespace AbbyyLS.SmartCAT.Selenium.Tests.Drivers
 	{
 		public static Logger Log = LogManager.GetCurrentClassLogger();
 
-		public RemoteWebDriver GetWebDriver(string tempFolder, string downloadDirectory)
+		public RemoteWebDriver GetWebDriver(string tempFolder, string downloadDirectory, string userDataDirectory)
 		{
 			var defaultDriverPath = new Uri(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "chromedriver.exe")).LocalPath;
 			var driverCopyPath = new Uri(Path.Combine(tempFolder, "chromedriver.exe")).LocalPath;
@@ -22,11 +22,17 @@ namespace AbbyyLS.SmartCAT.Selenium.Tests.Drivers
 
 			Directory.CreateDirectory(tempFolder);
 			Directory.CreateDirectory(downloadDirectory);
+			Directory.CreateDirectory(userDataDirectory);
 
-			File.Copy(defaultDriverPath, driverCopyPath);
+			if (!File.Exists(driverCopyPath))
+			{
+				File.Copy(defaultDriverPath, driverCopyPath);
+			}
 
 			var options = new ChromeOptions();
+			options.AddArgument("--enable-file-cookies");
 			options.AddArguments("--lang=en");
+			options.AddArguments(String.Format("--user-data-dir={0}", userDataDirectory));
 			options.AddUserProfilePreference("download.default_directory", downloadDirectory);
 			options.AddUserProfilePreference("safebrowsing.enabled", "true");
 			Log.Info("Настройки браузера {0}", String.Join(" ", options.Arguments));
